Align yagna with latest relay-client (#2688)
