+++ Cargo.toml
@@ -23,7 +23,7 @@ packet-trace-enable = [
     "ya-vpn/packet-trace-enable",
     "ya-file-logging/packet-trace-enable",
     "ya-net/packet-trace-enable",
-    "ya-service-bus/packet-trace-enable"
+    "ya-service-bus/packet-trace-enable",
 ]
 
 [[bin]]
@@ -215,7 +215,7 @@ members = [
     "utils/fd-metrics",
     "core/metrics",
     "test-utils/test-framework",
-    "test-utils/test-framework/framework-macro"
+    "test-utils/test-framework/framework-macro",
 ]
 
 [patch.crates-io]
@@ -244,10 +244,10 @@ ya-service-api-interfaces = { path = "core/serv-api/interfaces" }
 ya-service-api-web = { path = "core/serv-api/web" }
 
 ## SERVICE BUS
-ya-service-bus = { git = "https://github.com/golemfactory/ya-service-bus.git", rev = "190f0d772f7ed0830d54a2cef77d7a177f276c68"}
-ya-sb-proto = { git = "https://github.com/golemfactory/ya-service-bus.git", rev = "190f0d772f7ed0830d54a2cef77d7a177f276c68"}
-ya-sb-router = { git = "https://github.com/golemfactory/ya-service-bus.git", rev = "190f0d772f7ed0830d54a2cef77d7a177f276c68"}
-ya-sb-util = { git = "https://github.com/golemfactory/ya-service-bus.git", rev = "190f0d772f7ed0830d54a2cef77d7a177f276c68"}
+ya-service-bus = { git = "https://github.com/golemfactory/ya-service-bus.git", rev = "190f0d772f7ed0830d54a2cef77d7a177f276c68" }
+ya-sb-proto = { git = "https://github.com/golemfactory/ya-service-bus.git", rev = "190f0d772f7ed0830d54a2cef77d7a177f276c68" }
+ya-sb-router = { git = "https://github.com/golemfactory/ya-service-bus.git", rev = "190f0d772f7ed0830d54a2cef77d7a177f276c68" }
+ya-sb-util = { git = "https://github.com/golemfactory/ya-service-bus.git", rev = "190f0d772f7ed0830d54a2cef77d7a177f276c68" }
 
 #ya-service-bus = { path = "../ya-service-bus" }
 #ya-sb-proto = { path = "../ya-service-bus/crates/proto" }
@@ -259,11 +259,8 @@ ya-sb-util = { git = "https://github.com/golemfactory/ya-service-bus.git", rev =
 #ya-client-model = { git = "https://github.com/golemfactory/ya-client.git", rev = "2a6350f62cf8d926721225a3451822731456e3fe" }
 
 ## RELAY and networking stack
-ya-relay-stack = { git = "https://github.com/golemfactory/ya-relay.git", rev = "7d54d9b3421b6a2ed54d623210efb9e477920cdd" }
-ya-relay-client = { git = "https://github.com/golemfactory/ya-relay.git", rev = "7d54d9b3421b6a2ed54d623210efb9e477920cdd" }
-## TODO: We shouldn't reference `ya-relay-core` and `ya-relay-proto` directly
-ya-relay-core = { git = "https://github.com/golemfactory/ya-relay.git", rev = "7d54d9b3421b6a2ed54d623210efb9e477920cdd" }
-ya-relay-proto = { git = "https://github.com/golemfactory/ya-relay.git", rev = "7d54d9b3421b6a2ed54d623210efb9e477920cdd" }
+ya-relay-stack = { git = "https://github.com/golemfactory/ya-relay.git", rev = "ce4a2a97b75156a991b7e7836238ee8f60312508" }
+ya-relay-client = { git = "https://github.com/golemfactory/ya-relay.git", rev = "ce4a2a97b75156a991b7e7836238ee8f60312508" }
 
 #ya-relay-stack = { path = "../ya-relay/crates/stack" }
 #ya-relay-client = { path = "../ya-relay/client" }
