+++ core/net/Cargo.toml
@@ -11,12 +11,12 @@ service = []
 central-net = []
 packet-trace-enable = [
     "ya-packet-trace/enable",
-    "ya-relay-client/packet-trace-enable"
+    "ya-relay-client/packet-trace-enable",
 ]
 
 [dependencies]
 ya-client-model = "0.5"
-ya-core-model = { version = "^0.9", features=["net", "identity"] }
+ya-core-model = { version = "^0.9", features = ["net", "identity"] }
 
 ya-relay-client = "0.6"
 #ya-relay-client = { path = "../../../ya-relay/client" }
@@ -37,7 +37,7 @@ futures = "0.3"
 humantime = "2.1"
 lazy_static = "1.4"
 log = "0.4"
-metrics="0.12"
+metrics = "0.12"
 serde_json = "1.0"
 structopt = "0.3"
 strum = { version = "0.24", features = ["derive"] }
@@ -50,7 +50,7 @@ ethsign = { version = "0.8" }
 tokio-util = { version = "0.7" }
 url = { version = "2.2" }
 prost = { version = "0.10" }
-rand = { version = "0.7"}
+rand = { version = "0.7" }
 
 [dev-dependencies]
 ya-sb-proto = "0.6.1"
