--- core/net/src/hybrid_v2/mod.rs
@@ -1,11 +1,0 @@
-mod api;
-pub(crate) mod cli;
-mod client;
-mod codec;
-mod crypto;
-mod rest_api;
-mod service;
-
-pub use api::*;
-pub use rest_api::web_scope;
-pub use service::{send_bcast_new_neighbour, start_network, Net};
