+++ core/net/src/lib.rs
@@ -8,7 +8,6 @@ pub use service::{bind_broadcast_with_caller, broadcast, Net};
 mod bcast;
 pub mod central;
 pub mod hybrid;
-pub mod hybrid_v2;
 mod service;
 
 mod cli;
