+++ core/net/src/service.rs
@@ -57,10 +57,6 @@ impl Net {
                 crate::hybrid::cli::bind_service();
                 crate::hybrid::Net::gsb(ctx, config).await
             }
-            NetType::HybridV2 => {
-                crate::hybrid_v2::cli::bind_service();
-                crate::hybrid_v2::Net::gsb(ctx, config).await
-            }
         }
     }
 
@@ -69,7 +65,6 @@ impl Net {
         match net_type {
             NetType::Central => crate::central::web_scope(),
             NetType::Hybrid => crate::hybrid::web_scope(),
-            NetType::HybridV2 => crate::hybrid_v2::web_scope(),
         }
     }
 
@@ -83,7 +78,6 @@ impl Net {
         match &config.net_type {
             NetType::Central => Ok(()),
             NetType::Hybrid => crate::hybrid::Net::shutdown().await,
-            NetType::HybridV2 => crate::hybrid_v2::Net::shutdown().await,
         }
     }
 }
@@ -108,7 +102,6 @@ where
     match net_type {
         NetType::Central => crate::central::broadcast(caller, message).await,
         NetType::Hybrid => crate::hybrid::broadcast(caller, message).await,
-        NetType::HybridV2 => crate::hybrid_v2::broadcast(caller, message).await,
     }
 }
 
@@ -135,9 +128,6 @@ where
         }
         NetType::Hybrid => {
             crate::hybrid::bind_broadcast_with_caller(broadcast_address, handler).await
-        }
-        NetType::HybridV2 => {
-            crate::hybrid_v2::bind_broadcast_with_caller(broadcast_address, handler).await
         }
     }
 }
